import jsn log

+++ DB/SQL/GetInfo.sql
@@ -1,10 +1,18 @@
-﻿IF (OBJECT_ID('QueryStatInfo', 'V') IS NOT NULL) 
+﻿set QUOTED_IDENTIFIER ON;
+IF (OBJECT_ID('QueryStatInfo', 'V') IS NOT NULL) 
 	DROP VIEW QueryStatInfo;
+IF (OBJECT_ID('QueryStatInfo', 'U') IS NOT NULL) 
+	DROP TABLE QueryStatInfo;
 DROP VIEW IF EXISTS QueryStatInfo_Old;
 DROP VIEW IF EXISTS VwQueryStatInfo;
 DROP VIEW  IF EXISTS StatementStats;
 DROP VIEW IF EXISTS LockersStats;
 DROP VIEW IF EXISTS LockedStats;
+DROP VIEW  IF EXISTS VwStatementStats;
+DROP VIEW IF EXISTS VwLockersStats;
+DROP VIEW IF EXISTS VwLockedStats;
+DROP TABLE IF EXISTS Grouped_QueryStatInfo;
+DROP VIEW IF EXISTS VwQueryTextByDate;
 GO
 SET QUOTED_IDENTIFIER, ANSI_NULLS ON
 GO
@@ -28,7 +36,7 @@ INNER JOIN Dates d 	ON qh.DateId = d.Id
 INNER JOIN NormQueryTextHistory nqth ON qth.NormQueryTextHistoryId = nqth.Id
 GROUP BY qth.NormQueryTextHistoryId, nqth.NormalizedQuery, d.Date
 GO
-CREATE VIEW StatementStats WITH SCHEMABINDING AS 
+CREATE VIEW VwStatementStats WITH SCHEMABINDING AS 
 SELECT
 	qh.DateId
 	,qth.NormQueryTextHistoryId NormalizedQueryTextId
@@ -44,11 +52,11 @@ GROUP BY qh.DateId, qth.NormQueryTextHistoryId
 GO
 
 CREATE UNIQUE CLUSTERED INDEX Ix_Cl
-	ON StatementStats (DateId, NormalizedQueryTextId)
+	ON VwStatementStats (DateId, NormalizedQueryTextId)
 
-CREATE COLUMNSTORE INDEX Ix_AllColumns ON StatementStats (DateId, NormalizedQueryTextId, [Count], TotalDuration, TotalRowCount, TotalLogicalReads, TotalCPU, TotalWrites)
+CREATE COLUMNSTORE INDEX Ix_AllColumns ON VwStatementStats (DateId, NormalizedQueryTextId, [Count], TotalDuration, TotalRowCount, TotalLogicalReads, TotalCPU, TotalWrites)
 GO
-CREATE VIEW LockersStats WITH SCHEMABINDING
+CREATE VIEW VwLockersStats WITH SCHEMABINDING
 AS 
 SELECT li.DateId
 	  ,li.BlockerQueryId NormalizedQueryTextId
@@ -57,11 +65,11 @@ SELECT li.DateId
 FROM dbo.LongLocksInfo li
 GROUP BY li.DateId, li.BlockerQueryId
 GO
-CREATE UNIQUE CLUSTERED INDEX IX_Cl ON LockersStats (DateId, NormalizedQueryTextId);
-CREATE COLUMNSTORE INDEX Ix_AllColumns ON LockersStats (DateId, NormalizedQueryTextId, [Count], TotalDuration);
+CREATE UNIQUE CLUSTERED INDEX IX_Cl ON VwLockersStats (DateId, NormalizedQueryTextId);
+CREATE COLUMNSTORE INDEX Ix_AllColumns ON VwLockersStats (DateId, NormalizedQueryTextId, [Count], TotalDuration);
 GO
 
-CREATE VIEW LockedStats WITH SCHEMABINDING
+CREATE VIEW VwLockedStats WITH SCHEMABINDING
 AS 
 SELECT li.DateId
 	  ,li.BlockedQueryId NormalizedQueryTextId
@@ -70,14 +78,21 @@ SELECT li.DateId
 FROM dbo.LongLocksInfo li
 GROUP BY li.DateId, li.BlockedQueryId
 GO
-CREATE UNIQUE CLUSTERED INDEX IX_Cl ON LockedStats (DateId, NormalizedQueryTextId);
-CREATE COLUMNSTORE INDEX Ix_AllColumns ON LockedStats (DateId, NormalizedQueryTextId, [Count], TotalDuration);
+CREATE UNIQUE CLUSTERED INDEX IX_Cl ON VwLockedStats (DateId, NormalizedQueryTextId);
+CREATE COLUMNSTORE INDEX Ix_AllColumns ON VwLockedStats (DateId, NormalizedQueryTextId, [Count], TotalDuration);
 GO
-
+CREATE VIEW VwQueryTextByDate as 
+select nqth.Id TextId, 
+	d.id DateId
+FROM dbo.NormQueryTextHistory nqth
+CROSS JOIN dbo.Dates d
+go
+DROP VIEW IF EXISTS VwQueryStatInfo
+go
 CREATE VIEW VwQueryStatInfo
 AS
 SELECT
-	d.Date [Date]
+	qd.DateId [DateId]
    ,CAST(ROUND(s.TotalDuration, 2) AS DECIMAL(18, 2)) TotalDuration
    ,CAST(ROUND(s.TotalDuration / NULLIF(s.[count], 0), 2) AS DECIMAL(18, 2)) AvgDuration
    ,s.[count] [count]
@@ -86,52 +101,64 @@ SELECT
    ,CAST(ROUND(TotalCPU / NULLIF(s.[count], 0), 2) AS DECIMAL(18, 2)) AvgCPU
    ,CAST(ROUND(TotalWrites / NULLIF(s.[count], 0), 2) AS DECIMAL(18, 2)) AvgWrites
    ,CAST(0 AS DECIMAL(18, 2)) AS AvgAdoReads
-   ,nqth.NormalizedQuery QueryText
-   ,nqth.Id NormalizedQueryTextId
+   ,qd.TextId NormalizedQueryTextId
    ,lrs.[count] LockerCount
    ,CAST(ROUND(lrs.TotalDuration, 2) AS DECIMAL(18, 2)) LockerTotalDuration
    ,CAST(ROUND(lrs.TotalDuration / NULLIF(lrs.[count], 0), 2) AS DECIMAL(18, 2)) LockerAvgDuration
    ,lds.[count] LockedCount
    ,CAST(ROUND(lds.TotalDuration, 2) AS DECIMAL(18, 2)) LockedTotalDuration
    ,CAST(ROUND(lds.TotalDuration / NULLIF(lds.[count], 0), 2) AS DECIMAL(18, 2)) LockedAvgDuration
-   ,(SELECT COUNT_BIG(*) FROM DeadLocksInfo dli WHERE dli.DateId = d.Id AND (dli.QueryAId = nqth.Id OR dli.QuerybId = nqth.Id)) DeadLocksCount
-FROM dbo.NormQueryTextHistory nqth
-CROSS JOIN dbo.Dates d
-LEFT JOIN dbo.StatementStats s ON nqth.Id = s.NormalizedQueryTextId AND s.DateId = d.Id
-LEFT JOIN dbo.LockersStats lrs 	ON nqth.Id = lrs.NormalizedQueryTextId AND lrs.DateId = d.Id
-LEFT JOIN dbo.LockedStats lds 	ON nqth.Id = lds.NormalizedQueryTextId AND lds.DateId = d.Id
+   ,(SELECT COUNT_BIG(*) FROM DeadLocksInfo dli WHERE dli.DateId = qd.DateId AND (dli.QueryAId = qd.TextId OR dli.QuerybId = qd.TextId)) DeadLocksCount
+FROM VwQueryTextByDate qd
+LEFT JOIN dbo.VwStatementStats s ON qd.TextId = s.NormalizedQueryTextId AND s.DateId = qd.DateId
+LEFT JOIN dbo.VwLockersStats lrs ON qd.TextId = lrs.NormalizedQueryTextId AND lrs.DateId = qd.DateId
+LEFT JOIN dbo.VwLockedStats lds ON qd.TextId = lds.NormalizedQueryTextId AND lds.DateId = qd.DateId
+GO
 
+GO
+SELECT TOP 0 *
+INTO Grouped_QueryStatInfo
+FROM VwQueryStatInfo;
+
+CREATE CLUSTERED INDEX ix_cl ON Grouped_QueryStatInfo (DateId, NormalizedQueryTextId);
+CREATE COLUMNSTORE INDEX Ix_1 ON Grouped_QueryStatInfo (DateId, TotalDuration, AvgDuration, count, AvgRowCount
+, AvgLogicalReads, AvgCPU, AvgWrites, NormalizedQueryTextId,
+LockerCount, LockerTotalDuration, LockerAvgDuration, DeadLocksCount);
+GO
+
+CREATE VIEW QueryStatInfo AS
+SELECT gi.*
+	,t.NormalizedQuery QueryText
+	,d.[Date] [Date]
+FROM Grouped_QueryStatInfo gi
+INNER JOIN	 dbo.NormQueryTextHistory t ON t.Id = gi.NormalizedQueryTextId
+INNER JOIN	dbo.Dates d ON d.Id = gi.DateId
 GO
 
 DROP PROCEDURE  IF EXISTS ActualizeQueryStatInfo;
 GO
 
 CREATE PROCEDURE ActualizeQueryStatInfo AS BEGIN
+set QUOTED_IDENTIFIER ON;
+SET ANSI_NULLS ON;
 
 DECLARE @date DATETIME2 = DATEADD(DAY, -5, CAST(GETDATE() AS DATE));
-
-IF (OBJECT_ID('QueryStatInfo','U') IS NULL) BEGIN
-	SELECT TOP 0 *
-	INTO QueryStatInfo
-	FROM VwQueryStatInfo;
-
-	CREATE CLUSTERED INDEX ix_cl ON QueryStatInfo (Date, NormalizedQueryTextId);
-	CREATE COLUMNSTORE INDEX Ix_1 ON QueryStatInfo (Date, TotalDuration, AvgDuration, count, AvgRowCount
-	, AvgLogicalReads, AvgCPU, AvgWrites, NormalizedQueryTextId,
-	LockerCount, LockerTotalDuration, LockerAvgDuration, DeadLocksCount);
-
+IF NOT EXISTS (SELECT * FROM Grouped_QueryStatInfo) BEGIN
 	SET @date = CAST('2000-01-01' AS DATE);
 END ELSE BEGIN
-	DELETE QueryStatInfo
-	WHERE Date >= @date
+	DELETE Grouped_QueryStatInfo
+	FROM  dbo.Dates d
+	WHERE d.[Date] >= @date
+	AND Id = d.Id
 END;
 
 PRINT 'Date: ' + CAST(@date AS NVARCHAR(MAX));
 
-INSERT INTO QueryStatInfo
-SELECT *
-FROM VwQueryStatInfo
-WHERE [Date] >= @date
+INSERT INTO Grouped_QueryStatInfo
+SELECT v.*
+FROM VwQueryStatInfo v
+INNER JOIN dbo.Dates d ON d.Id = v.DateId
+WHERE d.[Date] >= @date;
 
 DBCC SHRINKFILE (N'work_analisys', 0, TRUNCATEONLY)
 DBCC SHRINKFILE (N'work_analisys_log', 0, TRUNCATEONLY)
